[52] - Test de users request test creado
